backend/requirements.txt
Flask==3.0.3
google-cloud-storage==2.18.2
pyodbc==5.1.0
google-auth==2.35.0
